# raylib: cache wrap text