About 10x speed up for html renderer's get total height (7ms to 0.7ms on device, 0.8ms to 0.08ms on PC)

still pretty slow, but way better and simpler than approach from https://github.com/commaai/openpilot/pull/36257